% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjts-package.R
\name{JTSVersion}
\alias{JTSVersion}
\title{Get JTS Version}
\usage{
JTSVersion()
}
\value{
character. Semantic version number of loaded JTS library.
}
\description{
Convenience method for accessing the version number for the currently-loaded JTS library.
}
\details{
When the JTS core library is loaded by rJava, either \code{.onAttach()} or after \code{install_jts()}, the \emph{org.locationtech.jts.JTSVersion} class static fields are read to determine the current version. This value is stored in the package option \code{rjts.JTS_VERSION}.
}
