% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjts-package.R
\docType{package}
\name{rjts-package}
\alias{rjts}
\alias{rjts-package}
\title{rjts: Interface to Java Topology Suite (JTS) Library}
\description{
The Java Topology Suite (JTS) is an open source library of spatial predicates and functions for processing geometries. \url{https://github.com/locationtech/jts} 'JTS' provides a complete, consistent, and robust implementation of fundamental algorithms for processing linear geometry on the 2-dimensional Cartesian plane. It also provides a comprehensive set of geometry test cases, and the 'TestBuilder' GUI application for working with and visualizing geometry and 'JTS' functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brownag/rjts}
  \item \url{https://humus.rocks/rjts}
}

}
\keyword{internal}
