% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{uninstall_jts}
\alias{uninstall_jts}
\title{Remove JAR files from user data directory}
\usage{
uninstall_jts(pattern = "\\\\.jar$")
}
\arguments{
\item{pattern}{Passed to \code{list.files()} REGEX pattern to match JAR files}
}
\value{
logical. result of recursive \code{all(file.remove(...))},  invisibly
}
\description{
Remove JAR files from user data directory
}
\examples{
\dontrun{
uninstall_jts()
}
}
