% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WKB.R
\name{WKBReader}
\alias{WKBReader}
\title{Create Well-known Binary Reader}
\usage{
WKBReader()
}
\value{
an object of class \code{jobjRef} to new instance of \code{org.locationtech.jts.io.WKBReader}
}
\description{
Creates a new reference to the JTS \code{WKBReader} class.
}
