% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WKT.R
\name{WKTReader}
\alias{WKTReader}
\title{Create Well-known Text Reader}
\usage{
WKTReader(...)
}
\arguments{
\item{...}{Optional: A custom \code{GeometryFactory}}
}
\value{
an object of class \code{jobjRef} to new instance of \code{org.locationtech.jts.io.WKTReader}
}
\description{
Creates a new reference to the JTS \code{WKTReader} class.
}
