% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WKT.R
\name{WKTWriter}
\alias{WKTWriter}
\title{Create Well-known Text Writer}
\usage{
WKTWriter()
}
\value{
an object of class \code{jobjRef} to new instance of \code{org.locationtech.jts.io.WKTWriter}
}
\description{
Creates a new reference to the JTS \code{WKTWriter} class.
}
