% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_jts}
\alias{install_jts}
\title{Install JTS Topology Suite from Latest GitHub Release}
\usage{
install_jts(
  path = find.package("rjts"),
  url = NULL,
  javadoc = FALSE,
  sources = FALSE,
  testbuilder = TRUE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{path}{Package folder to install into; Default: \code{find.package("rjts")}}

\item{url}{Optional: custom URL or path to JAR file}

\item{javadoc}{Download "Javadoc" JAR? Default: \code{FALSE}}

\item{sources}{Download "sources" JAR? Default: \code{FALSE}}

\item{testbuilder}{Download JTS Test Builder JAR? Default: \code{TRUE}}

\item{overwrite}{Overwrite existing JAR file (if present)? Default \code{FALSE}}

\item{...}{additional arguments passed to \code{utils::download.file()} (e.g. \code{quiet}, \code{method})}
}
\value{
Downloads and installs JAR files to \code{file.path(find.package("rjts"), "inst", "java")}, then calls \code{rJava::.jpackage()} to load the JAR files into Java classpath.
}
\description{
Downloads JTS JAR files to \code{inst/java} using \code{utils::download.file(..., mode="wb")}
}
\examples{
\dontrun{
install_jts()
}
}
