% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjts-package.R
\docType{package}
\name{rjts-package}
\alias{rjts}
\alias{rjts-package}
\title{rjts: R interface for the JTS Topology Suite}
\description{
The JTS Topology Suite \url{https://github.com/locationtech/jts} is a Java library for creating and manipulating vector geometry. It also provides a comprehensive set of geometry test cases, and the TestBuilder GUI application for working with and visualizing geometry and JTS functions.
}
\keyword{internal}
